Added error checking

+++ Midterm/src/com/util/jdbc/InitializeDatabase.java
@@ -22,7 +22,7 @@
 
  * Creation Date: 11, 2017 24
 
- * Last Modified: 11, 2017 25
+ * Last Modified: 11, 2017 29
 
  * Java Version: 1.8.1_141
 
@@ -36,6 +36,7 @@ package com.util.jdbc;
 
 import java.io.File;
 import java.io.FileNotFoundException;
+import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.Scanner;
 
@@ -50,27 +51,42 @@ public class InitializeDatabase {
 	
 	public InitializeDatabase(ConnectToDB dbConnection) {
 		
-		try {			
-			dbConnection.executeUpdate("DELETE FROM Players");
+		ResultSet rs;
+		try {
+			rs = dbConnection.executeQuerry("SELECT * FROM Players");
+			if(rs.next()) {
+				JOptionPane.showMessageDialog(null, "Database already exists. No action taken.", "Car Racing Game", javax.swing.JOptionPane.INFORMATION_MESSAGE);
+			}
+			else {
+				try {				
+					dbConnection.executeUpdate("DELETE FROM Players");
 
-			for(Scanner sc = new Scanner(new File("src\\com\\res\\data\\Prog32758Students.txt")); sc.hasNextLine(); ) {
-				
-				String line = sc.nextLine();
-				if (line.length() == 0) {
-					continue;
+					for(Scanner sc = new Scanner(new File("src\\com\\res\\data\\Prog32758Students.txt")); sc.hasNextLine(); ) {
+						
+						String line = sc.nextLine();
+						if (line.length() == 0) {
+							continue;
+						}
+						String[] tokens = line.split(",");
+			
+						dbConnection.executeUpdate("INSERT INTO Players (Last_Name, First_Name, `Group`) " + "VALUES ('" + tokens[0] + "', '" + tokens[1] + "', " + Integer.parseInt(tokens[2]) +")");
+						}
+					JOptionPane.showMessageDialog(null, "Database initialization successfully completed.\n Click OK to continue.", "Car Racing Game", JOptionPane.INFORMATION_MESSAGE);
 				}
-				String[] tokens = line.split(",");
-	
-				dbConnection.executeUpdate("INSERT INTO Players (Last_Name, First_Name, `Group`) " + "VALUES ('" + tokens[0] + "', '" + tokens[1] + "', " + Integer.parseInt(tokens[2]) +")");
+				catch (FileNotFoundException FileEx) {
+					JOptionPane.showMessageDialog(null, FileEx.getMessage() + " Error: File not found.", "Car Racing Game", javax.swing.JOptionPane.WARNING_MESSAGE);
 				}
-			JOptionPane.showMessageDialog(null, "Database initialization successfully completed.\n Click OK to continue.", "Car Racing Game", JOptionPane.INFORMATION_MESSAGE);
+				catch (SQLException SQLEx) {
+					JOptionPane.showMessageDialog(null, SQLEx.getMessage() + "SQL State: " + SQLEx.getSQLState() + " ErrorCode: " + SQLEx.getErrorCode(), "Car Racing Game", javax.swing.JOptionPane.WARNING_MESSAGE);
+				}
+			}
+		} catch (SQLException SQLEx) {
+			JOptionPane.showMessageDialog(null, "Database does not exist. Please select 'Create the Database' first.", "Car Racing Game", javax.swing.JOptionPane.WARNING_MESSAGE);
 		}
-		catch (FileNotFoundException ex) {
-			JOptionPane.showMessageDialog(null, ex.getMessage() + " Error: File not found.", "Car Racing Game", javax.swing.JOptionPane.WARNING_MESSAGE);
-		}
-		catch (SQLException e) {
-			JOptionPane.showMessageDialog(null, e.getMessage() + "SQL State: " + e.getSQLState() + " ErrorCode: " + e.getErrorCode(), "Car Racing Game", javax.swing.JOptionPane.WARNING_MESSAGE);
-		}
+		
+		
+		
+		
 	
 	}
